+++ src/main/java/com/example/lookup/repository/DescuentoRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DescuentoRepository extends JpaRepository<Descuento,Integer> {
-
+public interface DescuentoRepository extends JpaRepository<Descuento,Long> {
 }
